Move base64 related lisp-fns over to base64 module.
